Enable page scrolling when dragging things around
Adds a `Scroller` object when the browser is firefox using a crude
check against `navigator.userAgent`.

The `Scroller.scroll` method tries to detect whether user is currently
dragging something on the edge of the viewport. When a certain threshold
(hardcoded in variable `scrollSensitivity`) is reached, it tries to scroll
the page by `scrollSpeed` pixel.

Closes #19