Fix Sortable incorrectly disabled when handle is set
We do not know the initial state of the Sortable, so store it in the
`state` variable. When mouseout is triggered, mark Sortable as disabled
in the `state` variable.  The options hash still contains the original
value.

These two values are compared in the new `isDisabled()` method so that
`dragover` handler can then set the appropriate `draggable` attribute when
`handle` is set.

Closes #24